+++ Bundles/ProductQuantityStorage/src/Spryker/Client/ProductQuantityStorage/ProductQuantityStorageClientInterface.php
@@ -32,11 +32,11 @@ interface ProductQuantityStorageClientInterface
      * @api
      *
      * @param int $idProduct
-     * @param int $quantity
+     * @param float $quantity
      *
-     * @return int
+     * @return float
      */
-    public function getNearestQuantity(int $idProduct, int $quantity): int;
+    public function getNearestQuantity(int $idProduct, float $quantity): float;
 
     /**
      * Specification:
